requirements.txt
SpeechRecognition
pydub
python-docx
tqdm
